First commit
As usual my first commit is a big one.
The temperature controller app is responsible interfacing with the
thermostat on the kegerator.  The thermostat is made using an arduino,
and outputs the temperature inside the kegerator every 15 seconds.
The temperature data is recorded by the server and stored in the
database.